+++ src/main/java/com/rozsa/model/Item.java
@@ -7,26 +7,16 @@ import org.bson.types.ObjectId;
 
 public class Item implements Identifiable<ObjectId> {
     private String type;
-
     private ObjectId id;
-
     private long uid;
-
     private int level;
-
     private String name;
-
     private String description;
-
     private Sprite spriteData;
-
     private Icon iconData;
-
     private int count;
-
     private int maximumStack;
-
-    private int price;
+    private float price;
 
     public Item(String type) {
         this.type = type;
@@ -108,11 +98,11 @@ public class Item implements Identifiable<ObjectId> {
         this.maximumStack = maximumStack;
     }
 
-    public int getPrice() {
+    public float getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(float price) {
         this.price = price;
     }
